+++ src/mob_group.rs
@@ -16,7 +16,7 @@ use crate::SoundPlayer;
 pub struct MobGroup {
 
     // Group center x and y - the group should move as a whole
-    center: Vector2<f64>,
+    center: Vector2<f32>,
 
     members: Vec<MobGroupMember>,
 }
@@ -25,21 +25,21 @@ pub struct MobGroupMember {
     id: u64,
 
     // seconds till next action
-    action_countdown: f64,
+    action_countdown: f32,
     mobile: bool,
 }
 
 
 impl MobGroup {
 
-    pub fn new(mobs: Vec<u64>, center: Vector2<f64>, mobile: bool, rng: &mut StdRng) -> MobGroup {
+    pub fn new(mobs: Vec<u64>, center: Vector2<f32>, mobile: bool, rng: &mut StdRng) -> MobGroup {
 
         let mut members = Vec::with_capacity(mobs.len());
 
         for id in mobs {
             members.push(MobGroupMember {
                 id,
-                action_countdown: 0.1 + rng.random::<f64>(),
+                action_countdown: 0.1 + rng.random::<f32>(),
                 mobile,
             });
         }
@@ -51,7 +51,7 @@ impl MobGroup {
     }
 
 
-    pub fn update(&mut self, player_id: u64, dt: f64, mobs: &mut HashMap<u64, MapObject>, rng: &mut StdRng, 
+    pub fn update(&mut self, player_id: u64, dt: f32, mobs: &mut HashMap<u64, MapObject>, rng: &mut StdRng, 
                   factory: &mut MapObjectFactory, projectile_builder: &mut ProjectileBuilder,
                   speaker: &mut SoundPlayer) {
             
@@ -86,7 +86,7 @@ impl MobGroup {
                                 projectile_builder.configure_projectile("Iron shot", &mut projectile.visual, &mut projectile.velocity, speaker);
                                 mobs.insert(projectile.uid, projectile);
 
-                                member.action_countdown = 1.0 + rng.random::<f64>();
+                                member.action_countdown = 1.0 + rng.random::<f32>();
                             }
                         }
                         else if member.mobile {
@@ -97,8 +97,8 @@ impl MobGroup {
                             let mut y;
 
                             loop {
-                                x = mob.position[0] + 100.0 - rng.random::<f64>() * 200.0;
-                                y = mob.position[1] + 100.0 - rng.random::<f64>() * 200.0;
+                                x = mob.position[0] + 100.0 - rng.random::<f32>() * 200.0;
+                                y = mob.position[1] + 100.0 - rng.random::<f32>() * 200.0;
 
                                 let dx = x - self.center[0];
                                 let dy = y - self.center[1];
@@ -113,8 +113,8 @@ impl MobGroup {
 
                             if count >= 5 {
                                 println!("make {} return from {:?} to group center at {:?}", mob.uid, mob.position, self.center);
-                                x = self.center[0] + 50.0 - rng.random::<f64>() * 100.0;
-                                y = self.center[1] + 50.0 - rng.random::<f64>() * 100.0;
+                                x = self.center[0] + 50.0 - rng.random::<f32>() * 100.0;
+                                y = self.center[1] + 50.0 - rng.random::<f32>() * 100.0;
                             }
 
                             // println!("id=" + creature.id + "moves to " + x + ", " + y);
@@ -122,7 +122,7 @@ impl MobGroup {
                             let creature = mob.creature.as_ref().unwrap();
                             move_mob(mob, [x, y], creature.base_speed);
                             
-                            member.action_countdown = 3.0 + rng.random::<f64>() * 2.0;
+                            member.action_countdown = 3.0 + rng.random::<f32>() * 2.0;
                         }
                     }
                 }
